+++ rand_xoshiro/src/xoshiro256starstar.rs
@@ -73,6 +73,8 @@ impl Xoshiro256StarStar {
     }
 }
 
+impl_state_array_of_four!(Xoshiro256StarStar, u64);
+
 impl SeedableRng for Xoshiro256StarStar {
     type Seed = [u8; 32];
 
@@ -148,5 +150,12 @@ mod tests {
         let from_zero = Xoshiro256StarStar::from_seed([0u8; 32]);
         let from_sm0 = Xoshiro256StarStar::seed_from_u64(0);
         assert_eq!(from_zero, from_sm0);
+    }
+
+    #[test]
+    fn state_roundtrip() {
+        let rng = Xoshiro256StarStar::seed_from_u64(42);
+        let clone = Xoshiro256StarStar::from_seed(rng.state());
+        assert_eq!(clone, rng);
     }
 }